[Terrain] Fix refreshes on the Surface Gradient component. (#10093)
* Fix refreshes on the Surface Gradient component.
The Surface Gradient wasn't listening for surface changes, so it wasn't refreshing properly. However, it also couldn't listen for *all* surface changes, because that could cause feedback loops. For example, if the surface gradient is used by terrain, the gradient change could cause the terrain surface to change, which would trigger another gradient change, etc. The solution is to add the list of surface types affected to the OnSurfaceChanged message, so that it will ignore notifications for any surface tags except the ones that it specifically generates the gradient from.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* PR feedback and compile error on non-unity builds.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>