Cached projected shadows (#13288)
Adds support for shadows which will only update when they detect a change somewhere in their field of view. The option must be turned on inside the light component. Detecting that something moved is done as part of the view frustum cull, so the cost on the CPU side remains roughly the same as normal shadows, but in cases where the shadow doesn't need to re-render, the rendering cost is saved. This feature lends itself to scenes that are largely static like arch-vis.